+++ Sources/Features/Attributes/DefaultAttributes.swift
@@ -1,6 +1,13 @@
 // swiftlint:disable type_name
 import Foundation
 import CoreLocation
+#if os(iOS)
+import UIKit
+#elseif os(macOS)
+import AppKit
+#elseif os(tvOS)
+import UIKit
+#endif
 
 final class FaultInfo: AttributesSource {
     var faultMessage: String?
@@ -201,10 +208,17 @@ struct LibInfo: AttributesSource {
     private static let applicationLangName = "backtrace-cocoa"
 
     var backtraceVersion: String? {
+        
+#if SWIFT_PACKAGE
+        if let build = Bundle.main.infoDictionary?["CFBundleShortVersionString"] {
+            return build as? String
+        }
+#else
         if let bundle = Bundle(identifier: "Backtrace.io.Backtrace"),
            let build = bundle.infoDictionary?["CFBundleShortVersionString"] {
             return build as? String
         }
+#endif
         return nil
     }
 
@@ -277,6 +291,8 @@ private extension CLAuthorizationStatus {
         case .denied: return "Denied"
         case .notDetermined: return "notDetermined"
         case .restricted: return "restricted"
+        @unknown default:
+            return "Unknown"
         }
     }
 }
@@ -293,6 +309,8 @@ private extension UIDeviceOrientation {
         case .portrait: return "Portrait"
         case .portraitUpsideDown: return "PortraitUpsideDown"
         case .unknown: return "Unknown"
+        @unknown default:
+            return "Unknown"
         }
     }
 }
@@ -307,6 +325,8 @@ private extension UIDevice.BatteryState {
         case .full: return "Full"
         case .unknown: return "Unknown"
         case .unplugged: return "Unplugged"
+        @unknown default:
+            return "Unknown"
         }
     }
 }
@@ -320,6 +340,8 @@ private extension UIApplication.State {
         case .active: return "Active"
         case .background: return "Background"
         case .inactive: return "Inactive"
+        @unknown default:
+            return "Unknown"
         }
     }
 }