+++ Sources/Public/BacktraceCrashReporter.swift
@@ -1,5 +1,9 @@
 import Foundation
+#if SWIFT_PACKAGE
+import CrashReporter
+#else
 import Backtrace_PLCrashReporter
+#endif
 
 /// A wrapper around `PLCrashReporter`.
 @objc public class BacktraceCrashReporter: NSObject {
@@ -48,9 +52,8 @@ extension BacktraceCrashReporter: CrashReporting {
     func generateLiveReport(exception: NSException? = nil,
                             attributes: Attributes,
                             attachmentPaths: [String] = []) throws -> BacktraceReport {
-
-        let reportData = try reporter.generateLiveReport(with: exception)
-        return try BacktraceReport(report: reportData, attributes: attributes, attachmentPaths: attachmentPaths)
+        
+        return try BacktraceReport(report: reporter.generateLiveReport(), attributes: attributes, attachmentPaths: attachmentPaths)
     }
 
     func enableCrashReporting() throws {