+++ Tests/BacktraceFileManagerTests.swift
@@ -27,9 +27,13 @@ final class BacktraceFileManagerTests: QuickSpec {
                     }
                     throwingContext("given existing file") {
                         it("excludes file from backup") {
+#if SWIFT_PACKAGE
+                            guard let url = Bundle.module.url(forResource: "test", withExtension: "txt") else { fail(); return }
+#else
                             let bundle = Bundle(for: type(of: self))
                             guard let path = bundle.path(forResource: "test", ofType: "txt") else { fail(); return }
                             let url = URL(fileURLWithPath: path)
+#endif
                             expect {
                                 try BacktraceFileManager.excludeFromBackup(url)
                             }.toNot(throwError())
@@ -55,9 +59,19 @@ final class BacktraceFileManagerTests: QuickSpec {
                     }
                     throwingContext("given existing file") {
                         it("gets the size of a file") {
+
+                            
+                            
+#if SWIFT_PACKAGE
+                            guard let url = Bundle.module.url(forResource: "test", withExtension: "txt") else { fail(); return }
+#else
                             let bundle = Bundle(for: type(of: self))
                             guard let path = bundle.path(forResource: "test", ofType: "txt") else { fail(); return }
                             let url = URL(fileURLWithPath: path)
+#endif
+                            
+                            
+
                             expect {
                                 try BacktraceFileManager.sizeOfFile(at: url)
                             }.toNot(throwError())