+++ scripts/test.sh
@@ -7,5 +7,3 @@ set -o xtrace # to trace what gets executed. Useful for debugging
 fastlane ios tests
 fastlane mac tests
 fastlane tvos tests
-
-pod lib lint --verbose --allow-warnings --sources='https://cdn.cocoapods.org/'